% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_helpInput.R
\name{help_data}
\alias{help_data}
\title{modosin_data server function}
\usage{
help_data(input, output, session, modosindb, lang, main_data_reactives)
}
\arguments{
\item{input}{internal}

\item{output}{internal}

\item{session}{internal}

\item{lang}{lang reactive}
}
\description{
modosin_data server function
}
