% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_helpInput.R
\name{mod_helpInput}
\alias{mod_helpInput}
\title{mod_helpInput and siteDrought_data}
\usage{
mod_helpInput(id)
}
\arguments{
\item{id}{shiny id}
}
\description{
A shiny module to create and populate the data inputs
}
