% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_mapOutput_Polygon.R
\name{mod_mapOutput_polygon}
\alias{mod_mapOutput_polygon}
\title{mod_mapOutput and mod_map}
\usage{
mod_mapOutput_polygon(id)
}
\arguments{
\item{id}{shiny id}
}
\description{
Shiny module to generate the map
}
