% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_saveOutput.R
\name{mod_save}
\alias{mod_save}
\title{siteDrought_data server function}
\usage{
mod_save(input, output, session, main_data_reactives, data_reactives, lang)
}
\arguments{
\item{input}{internal}

\item{output}{internal}

\item{session}{internal}

\item{lang}{lang reactive}
}
\description{
siteDrought_data server function
}
