% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_saveOutput.R
\name{mod_saveOutput}
\alias{mod_saveOutput}
\title{mod_saveOutput and modosin_data}
\usage{
mod_saveOutput(id)
}
\arguments{
\item{id}{shiny id}
}
\description{
A shiny module to create and populate the data inputs
}
