% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_tsOutput.R
\name{mod_tsOutput}
\alias{mod_tsOutput}
\alias{mod_ts}
\title{mod_tsOutput and mod_ts}
\usage{
mod_tsOutput(id)

mod_ts(input, output, session, data_reactives, main_data_reactives, lang)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}

\item{data_reactives, main_data_reactives}{reactives needed}

\item{lang}{lang value}
}
\description{
A shiny module to generate and populate the time series output
}
