% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_dataInput_Polygon.R
\name{modosin_dataInput_polygon}
\alias{modosin_dataInput_polygon}
\title{modosin_dataInput  and modosin_data}
\usage{
modosin_dataInput_polygon(id)
}
\arguments{
\item{id}{shiny id}
}
\description{
A shiny module to create and populate the data inputs
}
