% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_dataInput.R
\name{siteDrought_data}
\alias{siteDrought_data}
\title{siteDrought_data server function}
\usage{
siteDrought_data(
  input,
  output,
  session,
  siteDroughtdb,
  lang,
  main_data_reactives
)
}
\arguments{
\item{input}{internal}

\item{output}{internal}

\item{session}{internal}

\item{lang}{lang reactive}
}
\description{
siteDrought_data server function
}
